move src/test_utils.ts to src/lib/
